+++ source/standalone_examples/replicator/offline_pose_generation/dope_pose_estimation.py
@@ -25,21 +25,14 @@ parser.add_argument("--max_queue_size", type=int, default=500, help="Max size of
 parser.add_argument("--output_folder", "-o", type=str, default="output", help="Output directory.")
 parser.add_argument("--use_s3", action="store_true", help="Saves output to s3 bucket. Only supported by DOPE writer.")
 parser.add_argument("--endpoint", type=str, default=None, help="s3 endpoint to write to.")
-parser.add_argument(
-    "--writer",
-    type=str,
-    default="YCBVideo",
-    help="Which writer to use to output data. Choose between: [YCBVideo, DOPE]",
-)
+
 args, unknown_args = parser.parse_known_args()
 
 if args.use_s3 and args.endpoint is None:
     raise Exception("To use s3, --endpoint must be specified.")
 
-CONFIG_FILES = {"dope": "dope_config.json", "ycbvideo": "ycb_config.json"}
-
 # Path to config file:
-CONFIG_FILE = CONFIG_FILES[args.writer.lower()]
+CONFIG_FILE = "dope_config.json"
 
 CONFIG_FILE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), "config", CONFIG_FILE)
 
@@ -62,6 +55,9 @@ F_X = config_data["F_X"]
 F_Y = config_data["F_Y"]
 C_X = config_data["C_X"]
 C_Y = config_data["C_Y"]
+
+# Default Camera Horizontal Aperture
+HORIZONTAL_APERTURE = config_data["HORIZONTAL_APERTURE"]
 
 # Number of sphere lights added to the scene
 NUM_LIGHTS = config_data["NUM_LIGHTS"]
@@ -109,11 +105,14 @@ kit = SimulationApp(launch_config=CONFIG)
 
 from omni.isaac.core.utils.stage import is_stage_loading
 from omni.isaac.core import World
-from omni.isaac.synthetic_utils import SyntheticDataHelper, DOPEWriter, YCBVideoWriter
+from omni.isaac.synthetic_utils import SyntheticDataHelper, DOPEWriter
 import omni.replicator.core as rep
 from omni.isaac.core.utils.nucleus import get_assets_root_path
 from omni.isaac.core.utils.semantics import add_update_semantics
 from omni.isaac.core.utils.rotations import euler_angles_to_quat
+from omni.isaac.core.utils.transformations import tf_matrix_from_pose, pose_from_tf_matrix
+from omni.kit.viewport_legacy import get_viewport_interface
+from omni.isaac.core.prims import XFormPrim
 from pxr import Usd, UsdGeom
 import math
 
@@ -125,21 +124,13 @@ from flying_distractors.dynamic_shape_set import DynamicShapeSet
 from flying_distractors.dynamic_object import DynamicObject
 from flying_distractors.dynamic_object_set import DynamicObjectSet
 from flying_distractors.flying_distractors import FlyingDistractors
-from camera_rig import CameraRig
-from utils import save_points_xyz, get_world_pose_from_relative
+from utils import get_world_pose_from_relative, get_random_world_pose_in_view, get_as_sdh_bb_3d_format
 
 
 class RandomScenario(torch.utils.data.IterableDataset):
-    def __init__(self, max_queue_size, num_mesh, num_dome, output_folder, use_s3=False, endpoint="", writer="dope"):
+    def __init__(self, max_queue_size, num_mesh, num_dome, output_folder, use_s3=False, endpoint=""):
 
         self.sd_helper = SyntheticDataHelper()
-
-        if writer == "dope":
-            self.writer_helper = DOPEWriter
-        elif writer == "ycbvideo":
-            self.writer_helper = YCBVideoWriter
-        else:
-            raise Exception("Invalid writer specified. Please ensure that writer is valid.")
 
         self.result = True
         assets_root_path = get_assets_root_path()
@@ -156,7 +147,7 @@ class RandomScenario(torch.utils.data.IterableDataset):
         self.train_part_mesh_path_to_prim_path_map = {}
         self.mesh_distractors = FlyingDistractors()
         self.dome_distractors = FlyingDistractors()
-        self.mesh = True
+        self.mesh_recording_active = True
 
         self.max_queue_size = max_queue_size
         self.data_writer = None
@@ -175,29 +166,46 @@ class RandomScenario(torch.utils.data.IterableDataset):
         signal.signal(signal.SIGINT, self._handle_exit)
 
     def _handle_exit(self, *args, **kwargs):
-
         print("exiting dataset generation...")
         self.exiting = True
 
     def _setup_world(self):
         """Populate scene with assets and prepare for synthetic data generation.
         """
-
         # Setup camera in simulation
-        self.camera_rig = CameraRig(
-            "/World/Rig",
-            "Rig",
-            "Camera",
-            WIDTH,
-            HEIGHT,
-            F_X,
-            F_Y,
-            C_X,
-            C_Y,
-            camera_rotation=CAMERA_ROTATION,
-            position=np.array([0, 0, -MAX_DISTANCE]),
-            orientation=euler_angles_to_quat(CAMERA_RIG_ROTATION),
+        focal_length = HORIZONTAL_APERTURE * F_X / WIDTH
+
+        # Setup camera and render product
+        self.camera = rep.create.camera(
+            position=(0, 0, -MAX_DISTANCE),
+            rotation=CAMERA_ROTATION,
+            focal_length=focal_length,
+            clipping_range=(0.01, 10000),
         )
+        self.render_product = rep.create.render_product(self.camera, (WIDTH, HEIGHT))
+
+        camera_node = self.camera.node
+        camera_xform_path = rep.utils.get_node_targets(camera_node, "inputs:prims")[0]
+        self.camera_path = str(camera_xform_path) + "/Camera"
+
+        # Define a prim relative to the camera with the camera's rotation "undone".
+        # Poses will be defined with respect to this prim.
+        camera_orientation = euler_angles_to_quat(CAMERA_ROTATION, degrees=True)
+        camera_transform = tf_matrix_from_pose(translation=(0, 0, 0), orientation=camera_orientation)
+        camera_transform_inv = np.linalg.inv(camera_transform)
+        _, rig_local_orientation = pose_from_tf_matrix(camera_transform_inv)
+
+        self.rig = XFormPrim(
+            prim_path=f"{self.camera_path}/rig", translation=np.array([0, 0, 0]), orientation=rig_local_orientation
+        )
+
+        # Create annotators
+        self.rgb = rep.AnnotatorRegistry.get_annotator("rgb")
+        self.rgb.attach(self.render_product)
+        self.bounding_box_3d = rep.AnnotatorRegistry.get_annotator("bounding_box_3d")
+        self.bounding_box_3d.attach(self.render_product)
+        self.occlusion = rep.AnnotatorRegistry.get_annotator("occlusion")
+        self.occlusion.attach(self.render_product)
 
         rep.settings.set_render_rtx_realtime()
 
@@ -207,17 +215,15 @@ class RandomScenario(torch.utils.data.IterableDataset):
         # Create a collision box in view of the camera, allowing distractors placed in the box to be within
         # [MIN_DISTANCE, MAX_DISTANCE] of the camera. The collision box will be placed in front of the camera,
         # regardless of CAMERA_ROTATION or CAMERA_RIG_ROTATION.
-        theta_x = self.camera_rig.fov_x / 2.0
-        theta_y = self.camera_rig.fov_y / 2.0
+        self.fov_x = 2 * math.atan(WIDTH / (2 * F_X))
+        self.fov_y = 2 * math.atan(HEIGHT / (2 * F_Y))
+        theta_x = self.fov_x / 2.0
+        theta_y = self.fov_y / 2.0
 
         # collision box dimensions lower than 1.3 do not work properly
         collision_box_width = max(2 * MAX_DISTANCE * math.tan(theta_x), 1.3)
         collision_box_height = max(2 * MAX_DISTANCE * math.tan(theta_y), 1.3)
         collision_box_depth = MAX_DISTANCE - MIN_DISTANCE
-
-        print(
-            f"collision_box_width: {collision_box_width} collision_box_height: {collision_box_height} collision_box_depth: {collision_box_depth} "
-        )
 
         collision_box_path = "/World/collision_box"
         collision_box_name = "collision_box"
@@ -230,9 +236,12 @@ class RandomScenario(torch.utils.data.IterableDataset):
         collision_box_rotation_from_camera = np.array([0, 0, 0])
         collision_box_orientation_from_camera = euler_angles_to_quat(collision_box_rotation_from_camera, degrees=True)
 
+        # Render a frame to ensure the Replicator Camera's underlying USD attributes are populated.
+        world.render()
+
         # Get the desired pose of the collision box from a pose defined locally with respect to the camera.
         collision_box_center, collision_box_orientation = get_world_pose_from_relative(
-            self.camera_rig.camera_path, collision_box_translation_from_camera, collision_box_orientation_from_camera
+            self.camera_path, collision_box_translation_from_camera, collision_box_orientation_from_camera
         )
 
         collision_box = CollisionBox(
@@ -360,10 +369,6 @@ class RandomScenario(torch.utils.data.IterableDataset):
         mesh_prim = world.stage.GetPrimAtPath(mesh_path)
         add_update_semantics(mesh_prim, prim_type)
 
-        if self.writer_helper == YCBVideoWriter:
-            # Save the vertices of the part in '.xyz' format. This will be used in one of PoseCNN's loss functions
-            save_points_xyz(path, mesh_path, prim_type, self._output_folder)
-
         # Add domain randomization with Omniverse Replicator Randomizers
         # Create and randomize sphere lights
         def randomize_sphere_lights():
@@ -375,7 +380,6 @@ class RandomScenario(torch.utils.data.IterableDataset):
                 scale=rep.distribution.uniform(1, 20),
                 count=NUM_LIGHTS,
             )
-
             return lights.node
 
         # Randomize prim colors
@@ -387,7 +391,6 @@ class RandomScenario(torch.utils.data.IterableDataset):
                 diffuse=rep.distribution.uniform((0, 0, 0), (1, 1, 1)),
                 count=100,
             )
-
             with prims:
                 rep.randomizer.materials(mats)
             return prims.node
@@ -409,49 +412,23 @@ class RandomScenario(torch.utils.data.IterableDataset):
 
         self.dome_distractors.set_visible(False)
 
-    def get_transform_matrices(self, bbox_data):
-        """Get transformation matrices for semantically labeled objects in view using bounding box data.
-
-        Args:
-            bbox_data (np.ndarray): Tight bounding box data. See get_bounding_box_2d_tight() in 
-                                    omni.syntheticdata.scripts.sensors for more details.
-
-        Returns:
-            np.ndarray: Column-major transformation matrices from frame of each visible prim to the world frame. Shape 
-                        is (num_visible_prims, 4, 4).
-        """
-        n = len(bbox_data)
-        if n > 0:
-            transform_matrices = np.zeros((n, 4, 4))
-        else:
-            return np.array([[[]]])
-
-        for i, bbox in enumerate(bbox_data):
-            mesh_path = bbox[1]
-            prim_path = self.train_part_mesh_path_to_prim_path_map[mesh_path]
-            prim = world.stage.GetPrimAtPath(prim_path)
-            prim_transform_matrix = UsdGeom.Xformable(prim).ComputeLocalToWorldTransform(Usd.TimeCode.Default())
-            prim_to_world = np.transpose(prim_transform_matrix)
-            transform_matrices[i, ...] = prim_to_world
-
-        return transform_matrices
-
     def randomize_movement_in_view(self, prim):
         """Randomly move and rotate prim such that it stays in view of camera.
 
         Args:
             prim (DynamicObject): prim to randomly move and rotate.
         """
-
-        translation, orientation = self.camera_rig.get_random_world_pose_in_view(
+        translation, orientation = get_random_world_pose_in_view(
+            self.camera_path,
             MIN_DISTANCE,
             MAX_DISTANCE,
+            self.fov_x,
+            self.fov_y,
             FRACTION_TO_SCREEN_EDGE,
-            self.camera_rig.prim_path,
+            self.rig.prim_path,
             MIN_ROTATION_RANGE,
             MAX_ROTATION_RANGE,
         )
-
         prim.set_world_pose(translation, orientation)
 
     def _capture_viewport(self):
@@ -474,65 +451,38 @@ class RandomScenario(torch.utils.data.IterableDataset):
             "OCCLUSION": {},
         }
 
-        if self.writer_helper == DOPEWriter:
-            gt_list = ["rgb", "boundingBox3D", "occlusion"]
-        elif self.writer_helper == YCBVideoWriter:
-            gt_list = ["rgb", "depthLinear", "boundingBox2DTight", "semanticSegmentation"]
-
-        # Collect Groundtruth
-        viewport = self.camera_rig.viewport_window
+        # Get viewport
+        viewport_interface = get_viewport_interface()
+        viewport_handle = viewport_interface.get_instance("Viewport")
+        viewport = viewport_interface.get_viewport_window(viewport_handle)
 
         # on the first frame make sure sensors are initialized
         if self.cur_idx == 0:
-            self.sd_helper.initialize(sensor_names=gt_list, viewport=viewport)
             kit.update()
             kit.update()
         # Render new frame
         kit.update()
 
-        # Collect Groundtruth
-        gt = self.sd_helper.get_groundtruth(gt_list, viewport)
-
         # RGB
-        image = gt["rgb"]
+        image = self.rgb.get_data()
         groundtruth["RGB"] = image
 
-        if self.writer_helper == YCBVideoWriter:
-            # Depth
-            groundtruth["DEPTH"] = gt["depthLinear"].squeeze()
+        # 3D BBox
+        repl_bb3d = self.bounding_box_3d.get_data()
+        groundtruth["BBOX3D"] = get_as_sdh_bb_3d_format(repl_bb3d)
 
-            # Semantic Segmentation
-            semantic_data = np.array(
-                self.sd_helper.get_mapped_semantic_data(gt["semanticSegmentation"], CLASS_NAME_TO_INDEX)
-            )
-            semantic_data[semantic_data == 65535] = 0  # deals with invalid semantic id
-            groundtruth["SEMANTIC"] = semantic_data
-
-            # 2D Tight BBox
-            groundtruth["BBOX2DTIGHT"] = gt["boundingBox2DTight"]
-
-            # Pose
-            groundtruth["POSE"]["PRIMSTOWORLD"] = self.get_transform_matrices(gt["boundingBox2DTight"])
-            groundtruth["POSE"]["CAMERAINTRINSICS"] = self.camera_rig.get_camera_intrinsic_matrix()
-
-        if self.writer_helper == DOPEWriter:
-            groundtruth["POSE"]["INDEXTOCLASSNAME"] = {
-                index: class_name for class_name, index in CLASS_NAME_TO_INDEX.items()
-            }
-            # 3D BBox
-            groundtruth["BBOX3D"] = gt["boundingBox3D"]
-
-            # Occlusion
-            groundtruth["OCCLUSION"] = gt["occlusion"]
+        # Occlusion
+        groundtruth["OCCLUSION"] = self.occlusion.get_data()
 
         # Pose
-        rig_transform = UsdGeom.Xformable(self.camera_rig.prim).ComputeLocalToWorldTransform(Usd.TimeCode.Default())
+        rig_transform = UsdGeom.Xformable(self.rig.prim).ComputeLocalToWorldTransform(Usd.TimeCode.Default())
         rig_to_world = np.transpose(rig_transform)
         groundtruth["POSE"]["DESIREDCAMERATOWORLD"] = rig_to_world
-        groundtruth["POSE"]["VIEWPARAMS"] = self.sd_helper.generic_helper_lib.get_view_params(
-            self.camera_rig.viewport_window
-        )
+        groundtruth["POSE"]["VIEWPARAMS"] = self.sd_helper.generic_helper_lib.get_view_params(viewport)
         groundtruth["POSE"]["CLASSNAMETOINDEX"] = CLASS_NAME_TO_INDEX
+        groundtruth["POSE"]["INDEXTOCLASSNAME"] = {
+            index: class_name for class_name, index in CLASS_NAME_TO_INDEX.items()
+        }
 
         self.data_writer.q.put(groundtruth)
         return image
@@ -541,10 +491,9 @@ class RandomScenario(torch.utils.data.IterableDataset):
         return self
 
     def __next__(self):
-
         if self.cur_idx == self.num_mesh:  # MESH datset generation complete, switch to DOME dataset
 
-            self.mesh = False
+            self.mesh_recording_active = False
 
             # Hide the FlyingDistractors used for the MESH dataset
             self.mesh_distractors.set_visible(False)
@@ -571,7 +520,7 @@ class RandomScenario(torch.utils.data.IterableDataset):
 
             world.step(render=False)
 
-        if self.mesh:
+        if self.mesh_recording_active:
             flying_distractors = self.mesh_distractors
         else:
             flying_distractors = self.dome_distractors
@@ -591,21 +540,15 @@ class RandomScenario(torch.utils.data.IterableDataset):
 
         # Write to disk
         if self.data_writer is None:
-            if self.writer_helper == DOPEWriter:
-                self.data_writer = self.writer_helper(
-                    data_dir=self._output_folder,
-                    num_worker_threads=self._num_worker_threads,
-                    num_frames=self.train_size,
-                    max_queue_size=self.max_queue_size,
-                    use_s3=self.use_s3,
-                    endpoint_url=self.endpoint,
-                    bucket_name=self._output_folder,
-                )
-            elif self.writer_helper == YCBVideoWriter:
-                self.data_writer = self.writer_helper(
-                    self._output_folder, self._num_worker_threads, self.train_size, self.max_queue_size
-                )
-
+            self.data_writer = DOPEWriter(
+                data_dir=self._output_folder,
+                num_worker_threads=self._num_worker_threads,
+                num_frames=self.train_size,
+                max_queue_size=self.max_queue_size,
+                use_s3=self.use_s3,
+                endpoint_url=self.endpoint,
+                bucket_name=self._output_folder,
+            )
             self.data_writer.start_threads()
 
         image = self._capture_viewport()
@@ -620,7 +563,6 @@ dataset = RandomScenario(
     output_folder=args.output_folder,
     use_s3=args.use_s3,
     endpoint=args.endpoint,
-    writer=args.writer.lower(),
 )
 
 num_frames = args.num_mesh + args.num_dome
@@ -649,6 +591,8 @@ if dataset.result:
 
     print("end:", datetime.datetime.now().strftime("%m/%d/%Y, %H:%M:%S"))
     print("Total time taken: ", str(datetime.datetime.now() - start_time).split(".")[0])
+
+kit.update()
 
 # cleanup
 kit.close()